k3sCluster/scripts/requirements.txt
pandas
pathlib
pyyaml